renaming components and style
