+++ src/components/eyubkh/Button.jsx
@@ -1,4 +1,4 @@
-export const GenerateButton = ({ handler }) => {
+export const Button = ({ handler }) => {
 	return (
 		<button id='eyubkh-button' onClick={handler}>
 			Generate
