--- src/components/eyubkh/Home.jsx
@@ -1,11 +1,0 @@
-import { GeneratePassword } from './GeneratePassword'
-import { Animation } from './Animation'
-
-export default function Home() {
-	return (
-		<>
-			<GeneratePassword />
-			<Animation />
-		</>
-	)
-}
