+++ src/components/eyubkh/Page.jsx
@@ -1,0 +1,13 @@
+import { Card } from './Card'
+import { Castle } from './Castle'
+
+import './style/global.css'
+
+export default function Page() {
+	return (
+		<>
+			<Card />
+			<Castle />
+		</>
+	)
+}
