+++ src/components/eyubkh/style/card.css
@@ -2,6 +2,7 @@
 	display: flex;
 	flex-direction: column;
 	position: absolute;
+	width: 420px;
 	gap: 10px;
 	left: 50%;
 	top: 10%;
