+++ src/components/eyubkh/style/global.css
@@ -1,0 +1,5 @@
+@media only screen and (max-width: 720px) {
+	body {
+		background-color: aqua;
+	}
+}
